Style: 공연 관련 페이지 라우트 변경
